change user upload

+++ src/controllers/user.controller.ts
@@ -2,6 +2,8 @@ import { Request, Response } from "express";
 import prisma from "../configs/prisma.config";
 import fs from "node:fs"
 import path from "node:path";
+import { getUserImage } from "../services/image";
+import { users } from "@prisma/client";
 
 const getUser = async (req: Request, res: Response) => {
   const { user_level } = res.locals.token;
@@ -47,49 +49,45 @@ const updateUser = async (req: Request, res: Response) => {
     res.status(400).json({ status: 400 ,message: "ไม่พบไฟล์รูป" })
   }else{
     const { user_name, display_name, user_status } = req.body;
-    const { user_id, filename } = req.params;
+    const { user_id } = req.params;
     let file: string = req.file?.filename; 
-    let value = {
-      user_name: user_name,
-      display_name: display_name,
-      user_status: Number(user_status), 
-      user_picture: `/images/${file}` 
-    };
-    await prisma.users.update({
-      where: { 
-        user_id : user_id 
-      },
-      data: value
-    }).then((result) => {
-      fs.unlinkSync(path.join('public/images', filename));
+    try{
+      let value = {
+        user_name: user_name,
+        display_name: display_name,
+        user_status: Number(user_status), 
+        user_picture: `/images/${file}` 
+      };
+      const filename = await getUserImage(user_id);
+      const result: users = await prisma.users.update({
+        where: { 
+          user_id : user_id 
+        },
+        data: value
+      })
+      fs.unlinkSync(path.join('public/images', String(filename?.split("/")[2])));
       res.json({ 
         status: 200,
         msg: 'Update Successful!!',
         value : result
       });
-    }).catch((err) => {
+    }catch(err){
       fs.unlinkSync(path.join('public/images', file));
       res.status(400).json({ error: err });
-    });
+    }
   }
 }
 
 const deleteUser = async (req: Request, res: Response) => {
-  const { user_id, filename } = req.params;
-  await prisma.users.delete({
-    where: { 
-      user_id: user_id 
-    }
-  }).then((result) => {
-    fs.unlinkSync(path.join('public/images', filename));
-    res.json({ 
-      status: 200,
-      msg: 'Delete Successful!!',
-      value : result
-    });
-  }).catch((err) => {
+  const { user_id } = req.params;
+  try{
+    const filename = await getUserImage(user_id);
+    await prisma.users.delete({ where: { user_id: user_id } })
+    fs.unlinkSync(path.join('public/images', String(filename?.split("/")[2])));
+    res.json({ status: 200, msg: 'Delete Successful!!' });
+  }catch(err){
     res.status(400).json({ error: err });
-  });
+  }
 }
 
 export {
